+++ utils/utils.go
@@ -1,7 +1,11 @@
 package utils
 
-import "log"
-import "net"
+import (
+	"io/ioutil"
+	"log"
+	"net"
+	"net/http"
+)
 
 // BuildAppioidBaseURL detects ip and concat with port to URL
 func BuildAppioidBaseURL(port string) string {
@@ -18,4 +22,15 @@ func BuildAppioidBaseURL(port string) string {
 // AppiumServerURL build URL for interact with appium
 func AppiumServerURL(port string) string {
 	return "http://127.0.0.1:" + port + "/wd/hub"
+}
+
+// AppiumStatus return current atatus of Appium server
+func AppiumStatus(port string) string {
+	response, err := http.Get(AppiumServerURL(port) + "/status")
+	if err == nil {
+		defer response.Body.Close()
+		responseData, _ := ioutil.ReadAll(response.Body)
+		return string(responseData)
+	}
+	return "ERR"
 }
